Use unique pointers instead of copying values when creating nodes
The previous implementation was not using smart pointers which resulted
in a lot of copies. For example when a child node was added to the
vector of the children of its parent, the node itself and its vector of
children and all its descendants were copied. This is an inefficient
process! Maybe the compiler optimized that but nothing ensure that.

We now use smart pointers so the nodes are created only once and only
the smart pointer ownership is moved. It reduces a lot of time spent in
copying data.

FossilOrigin-Name: ca9ad9fdfbc064af60da21ea0b2cd232ad241b788f59c38faf4074bdfb5753c4